macos - EUREKA - fix runtime fsharp.core.dll versioning conflicts
The issue was that even though the F# library authoring guide here: https://fsharp.github.io/2015/04/18/fsharp-core-notes.html#do-not-bundle-fsharpcore-with-a-library says to use <PackageReference /> in fsproj to pin FSharp.Core down to a lower version when writing a library, you cannot do that when using paket, as the PackageReference element is specifically NuGet. Thus, your library will build against a different version of FSharp.Core than it was shipped with. In my case here, it got an older FSharp.Core from paket at runtime, causing all sorts of difficult to diagnose runtime weirdness (specifically about not being able to locate async types).